move sprite types to sprites.ts, remove redundant idle positioning, clean up
